New UI design

+++ client/src/components/dashboard/Experience.js
@@ -4,11 +4,12 @@ import PropTypes from "prop-types";
 import Moment from "react-moment";
 import { deleteExperience } from "../../store/actions/profileActions";
 
+// svg
+import trash from "../../img/SVG/trash.svg";
+
 class Experience extends Component {
   handleClick = id => {
-    if (window.confirm("Are You Sure ?")) this.props.deleteEducation(id);
-
-    this.props.deleteExperience(id);
+    if (window.confirm("Are You Sure ?")) this.props.deleteExperience(id);
   };
 
   render() {
@@ -16,43 +17,49 @@ class Experience extends Component {
 
     if (this.props.experience) {
       experience = this.props.experience.map(exp => (
-        <tr key={exp._id}>
-          <td>{exp.company}</td>
-          <td>{exp.title}</td>
-          <td>
-            <Moment format="YYYY/MM/DD">{exp.from}</Moment> -{" "}
-            {exp.to === null ? (
-              "Now"
-            ) : (
-              <Moment format="YYYY/MM/DD">{exp.to}</Moment>
-            )}
-          </td>
-          <td>
-            {" "}
-            <button
-              onClick={this.handleClick.bind(this, exp._id)}
-              className="btn btn-danger"
-            >
-              &times;
-            </button>{" "}
-          </td>
-        </tr>
+        <div className="credentials__card" key={exp._id}>
+          <ul className="credentials__card--details">
+            <li className="credentials__card--details--item">
+              <span className="credentials--span">Company: </span> {exp.company}
+            </li>
+            <li className="credentials__card--details--item">
+              <span className="credentials--span">Title: </span> {exp.title}
+            </li>
+            <li className="credentials__card--details--item">
+              <span className="credentials--span">Years: </span>{" "}
+              <Moment format="MM-DD-YYYY">{exp.from}</Moment> -{" "}
+              {exp.to === null ? (
+                "Till Present"
+              ) : (
+                <Moment format="YYYY-MM-DD">{exp.to}</Moment>
+              )}
+            </li>
+            <li className="credentials__card--details--item">
+              {exp.description === "" ? null : (
+                <span>
+                  <span className="credentials--span">Description: </span>
+                  <span>{exp.description}</span>
+                </span>
+              )}
+            </li>
+          </ul>
+          <button
+            className="credentials__delete-btn"
+            onClick={this.handleClick.bind(this, exp._id)}
+          >
+            <img
+              className="credentials__icon-delete"
+              src={trash}
+              alt="delete button"
+            />
+          </button>
+        </div>
       ));
     }
     return (
-      <div>
-        <h4 className="mb-4">Experience Credentials</h4>
-        <table className="table">
-          <tbody>
-            <tr>
-              <th>Company</th>
-              <th>Title</th>
-              <th>Years</th>
-              <th />
-            </tr>
-            {experience}
-          </tbody>
-        </table>
+      <div className="credentials-section">
+        <h4 className="heading-tertiary--main">Experience Credentials</h4>
+        <div className="credentials">{experience}</div>
       </div>
     );
   }
